requirements.txt
PyYAML~=5.4.1
numpy~=1.22.3
pandas~=1.4.2
sklearn~=0.0
scikit-learn~=1.0.2
bottle~=0.12.19
sphinx